GUI Rendering: exports image, shows it when done
A few to-do items:
- Add progress messages to the GUI, such as "preparing" while the render
memory is being allocated, and "saving image" while the image is being
exported, as both tend to take more than a few seconds when the image is
sufficiently large.
- Don't ever clone a Render struct, because it holds ALL the pixel data,
which is a ton of memory, and images larger than about 2k pixels are
exhausting the memory and crashing the program.
- In the GUI, add a feature to save the image (which will just require
copying the exported file from the /tmp directory to a user-decided
location
- Maybe add a "Cancel" button while the render is in progress, which
halts the device threads, or whatever is happening at that time